sys: kern: arch: rv32: move atomic bool to its own module.
